Adding mongodb local env config + travis integration. Refactoring api.js into storyController. Adding new gulp test task.
