+++ .travis.yml
@@ -1,6 +1,8 @@
 language: node_js
 node_js:
   - "4.2"
+services:
+  - mongodb
 install:
   - npm install
 script:
